Versao funcional do codigo otimizado

+++ src/main.c
@@ -8,30 +8,30 @@
 
 int main() {
     int n, max_iter; // define  o tamanho do sistema e o número máximo de iterações
-    double x0, tol; // define o chute inicial e a tolerância para convergência
+    long double x0, tol; // define o chute inicial e a tolerância para convergência
 
     printf ("Digite a dimensao do sistema (N): ");
     scanf ("%d", &n);
     printf ("Digite o chute inicial (x0): ");
-    scanf ("%lf", &x0);
+    scanf ("%Lf", &x0);
     printf ("Digite a tolerancia (TOL): ");
-    scanf ("%lf", &tol);
+    scanf ("%Lf", &tol);
     printf ("Digite o numero maximo de iteracoes: ");
     scanf ("%d", &max_iter);
 
-    double *x =  malloc(n * sizeof(double)); // vetor com as variáveis x
-    double *f =  malloc(n * sizeof(double)); // vetor com cada resultado de f(x)
-    double *s =  malloc(n * sizeof(double)); // delta usado para atualizar x (i+1)
+    long double *x =  malloc(n * sizeof(long double)); // vetor com as variáveis x
+    long double *f =  malloc(n * sizeof(long double)); // vetor com cada resultado de f(x)
+    long double *s =  malloc(n * sizeof(long double)); // delta usado para atualizar x (i+1)
 
-    // double **j =  malloc(n * sizeof(double*)); // matriz jacobiana
+    // long double **j =  malloc(n * sizeof(long double*)); // matriz jacobiana
     // for (int i = 0; i < n; i++){
-    //     j[i] = malloc(n * sizeof(double));
+    //     j[i] = malloc(n * sizeof(long double));
     // }
     
     // Jacobiana otimizada
-    double *a =  malloc(n * sizeof(double)); // subdiagonal
-    double *b =  malloc(n * sizeof(double)); // diagonal principal
-    double *c =  malloc(n * sizeof(double)); // superdiagonal
+    long double *a =  malloc(n * sizeof(long double)); // subdiagonal
+    long double *b =  malloc(n * sizeof(long double)); // diagonal principal
+    long double *c =  malloc(n * sizeof(long double)); // superdiagonal
     
 
     /* chute inicial para todos */
